product image handling mostly done now

+++ database/migrations/2016_08_31_022917_create_product_type_images_table.php
@@ -18,7 +18,7 @@ class CreateProductTypeImagesTable extends Migration
             $table->string('filename');
             $table->integer('width')->nullable();
             $table->integer('height')->nullable();
-            $table->float('order')->nullable();
+            $table->integer('order')->nullable();
             $table->timestamps();
         });
     }
